+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## Unreleased
 
-（なし）
+- eframe: persistence feature flagを追加
+- generic: デフォルトでエラー時にビープ音を鳴らすように
 
 ### デモプラグイン
 
-（なし）
+- image-rs-input: アルファ値周りの処理を修正
 
 ## [0.17.0](https://github.com/sevenc-nanashi/aviutl2-rs/releases/tag/0.17.0)